Dump
Dump of all the files related to the project "Global sampling of Feynman diagrams through Normalizing Flow".
The code has been refactored to make it suitable for pubblication, but has not been retested right now since needs a cuda machine to compile the ML part.
Right now the CPU part realted to the DMC library compiles fine, in the next days tests will be performed on the NF module.